+++ src/__tests__/unit/disableAll.test.ts
@@ -1,52 +1,71 @@
-import { setupTest } from "../../utils/testHelper";
+import "jest-to-log";
+import { NodeLogger } from "../../index";
+import { expect } from "@jest/globals";
 
 describe("disableAll() method", () => {
-  const { createWorkerDataPromise } = setupTest(
-    "unit",
-    "disableAll.test.process.ts"
-  );
-
-  /**
-   *
-   * Only createWorkerDataPromise; Nu createMessagePromise
-   * @returns {Promise<{discolouredResult: string}>}
-   */
-  async function processWorkerPromises(): Promise<{
-    discolouredResult: string;
-  }> {
-    const promiseResult = await createWorkerDataPromise();
-    const workerData = promiseResult;
-
-    const discolouredResult = workerData.replace(
-      /(\x1b\[\d+;\d+m)|(\x1b\[\d+m)/g,
-      ""
-    );
-
-    return { discolouredResult };
-  }
-
   describe("1) given that disableAll() is called directly on an instance", () => {
-    it("should log nothing", async () => {
-      const { discolouredResult } = await processWorkerPromises();
+    it("should log nothing", () => {
+      const logger = new NodeLogger("Log tester");
 
-      //just \n
-      const expected = "\n";
+      let logResult = {};
 
-      expect(discolouredResult).toBe(expected);
+      function testFn() {
+        logger.disableAll();
+
+        const foo = "bar";
+        const logReturn = logger.log({ foo });
+
+        logResult = { ...logReturn };
+      }
+
+      expect(testFn).toLogStdout("");
+
+      //logResult assigned only after the abv assertion
+      expect(logResult).toStrictEqual({
+        stack: null,
+        logTitle: null,
+        logBody: null,
+      });
     });
   });
 
   describe("2) given that disableAll() is called on the constructor (indirectly on the instance)", () => {
-    it("should log nothing", async () => {
-      const { discolouredResult } = await processWorkerPromises();
+    it("should log nothing", () => {
+      const logger = new NodeLogger("Log tester").disableAll();
 
-      const expected = "\n";
-      expect(discolouredResult).toBe(expected);
+      let logResult = {};
+
+      function testFn() {
+        const foo = "bar";
+        const logReturn = logger.log({ foo });
+        logResult = { ...logReturn };
+      }
+      expect(testFn).toLogStdout("");
+
+      expect(logResult).toStrictEqual({
+        stack: null,
+        logTitle: null,
+        logBody: null,
+      });
     });
   });
-  describe("3) given that the log method is called after invoking disableAll() on the constructor", () => {
-    it("should NOT throw", async () => {
-      expect(async () => await processWorkerPromises()).not.toThrow();
+  describe("3) given that the log method is called by the instance after invoking disableAll() on the constructor", () => {
+    it("should NOT throw", () => {
+      const logger = new NodeLogger("Log tester").disableAll();
+      let logResult = {};
+
+      function testFn() {
+        const foo = "bar";
+        const logReturn = logger.log({ foo });
+        logResult = { ...logReturn };
+      }
+
+      expect(testFn).toLogStdout("");
+      expect(logResult).toStrictEqual({
+        stack: null,
+        logTitle: null,
+        logBody: null,
+      });
     });
   });
 });
